Break out sequence.hpp
